Moved Linux specific events
