+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/arch/linux/event/cond/FenceCond.java
@@ -1,24 +1,21 @@
-package com.dat3m.dartagnan.program.event.rmw.cond;
+package com.dat3m.dartagnan.program.arch.linux.event.cond;
 
+import com.dat3m.dartagnan.program.event.Event;
+import com.dat3m.dartagnan.program.event.Fence;
 import com.dat3m.dartagnan.utils.recursion.RecursiveAction;
 import com.dat3m.dartagnan.verification.VerificationTask;
-
 import org.sosy_lab.java_smt.api.BooleanFormula;
 import org.sosy_lab.java_smt.api.BooleanFormulaManager;
 import org.sosy_lab.java_smt.api.SolverContext;
 
-import com.dat3m.dartagnan.expression.ExprInterface;
-import com.dat3m.dartagnan.expression.IExpr;
-import com.dat3m.dartagnan.program.event.Event;
-import com.dat3m.dartagnan.program.event.rmw.RMWStore;
-import com.dat3m.dartagnan.program.event.utils.RegReaderData;
+public class FenceCond extends Fence {
 
-public class RMWStoreCond extends RMWStore implements RegReaderData {
-
+    private final RMWReadCond loadEvent;
     protected transient BooleanFormula execVar;
 
-    public RMWStoreCond(RMWReadCond loadEvent, IExpr address, ExprInterface value, String mo) {
-        super(loadEvent, address, value, mo);
+    public FenceCond (RMWReadCond loadEvent, String name){
+        super(name);
+        this.loadEvent = loadEvent;
     }
 
     @Override
@@ -34,20 +31,21 @@ public class RMWStoreCond extends RMWStore implements RegReaderData {
 
     @Override
     public String toString(){
-        return String.format("%1$-" + Event.PRINT_PAD_EXTRA + "s", super.toString()) + ((RMWReadCond)loadEvent).condToString();
+        return String.format("%1$-" + Event.PRINT_PAD_EXTRA + "s", super.toString()) + loadEvent.condToString();
     }
 
     @Override
     protected BooleanFormula encodeExec(SolverContext ctx){
         BooleanFormulaManager bmgr = ctx.getFormulaManager().getBooleanFormulaManager();
-		return bmgr.equivalence(execVar, bmgr.and(cfVar, ((RMWReadCond)loadEvent).getCond()));
+		return bmgr.equivalence(execVar, bmgr.and(cfVar, loadEvent.getCond()));
     }
 
     // Unrolling
     // -----------------------------------------------------------------------------------------------------------------
 
+
     @Override
     public RecursiveAction unrollRecursive(int bound, Event predecessor, int depth) {
-        throw new RuntimeException("RMWStoreCond cannot be unrolled: event must be generated during compilation");
+        throw new RuntimeException("FenceCond cannot be unrolled: event must be generated during compilation");
     }
 }
