+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/arch/linux/event/cond/RMWReadCondCmp.java
@@ -1,4 +1,4 @@
-package com.dat3m.dartagnan.program.event.rmw.cond;
+package com.dat3m.dartagnan.program.arch.linux.event.cond;
 
 import com.dat3m.dartagnan.expression.ExprInterface;
 import com.dat3m.dartagnan.expression.IExpr;
